JMX增强
git-svn-id: http://code.alibabatech.com/svn/druid/trunk@705 b9813039-fb51-4c41-a8b9-e21c2acb5095

+++ src/main/java/com/alibaba/druid/pool/DruidAbstractDataSource.java
@@ -143,7 +143,7 @@ public abstract class DruidAbstractDataSource implements DruidAbstractDataSource
 
     protected boolean                                                                        removeAbandoned;
 
-    protected int                                                                            removeAbandonedTimeoutMillis;
+    protected long                                                                           removeAbandonedTimeoutMillis;
 
     protected boolean                                                                        logAbandoned;
 
@@ -283,18 +283,18 @@ public abstract class DruidAbstractDataSource implements DruidAbstractDataSource
     }
 
     public int getRemoveAbandonedTimeout() {
-        return removeAbandonedTimeoutMillis / 1000;
+        return (int) (removeAbandonedTimeoutMillis / 1000);
     }
 
     public void setRemoveAbandonedTimeout(int removeAbandonedTimeout) {
         this.removeAbandonedTimeoutMillis = removeAbandonedTimeout * 1000;
     }
 
-    public void setRemoveAbandonedTimeoutMillis(int removeAbandonedTimeoutMillis) {
+    public void setRemoveAbandonedTimeoutMillis(long removeAbandonedTimeoutMillis) {
         this.removeAbandonedTimeoutMillis = removeAbandonedTimeoutMillis;
     }
 
-    public int getRemoveAbandonedTimeoutMillis() {
+    public long getRemoveAbandonedTimeoutMillis() {
         return removeAbandonedTimeoutMillis;
     }
